%s+% {stringi}	R Documentation

Concatenate Two Character Vectors

Description

Binary operators for joining (concatenating) two character vectors, with a typical R look-and-feel.

Usage

e1 %s+% e2

e1 %stri+% e2

Arguments

e1
a character vector or an object coercible to a character vector

e2
a character vector or an object coercible to a character vector

Details

Vectorized over e1 and e2.

These operators act like a call to stri_join(e1, e2, sep=""). However, note that joining 3 vectors, e.g. e1 %+% e2 %+% e3 is slower than stri_join(e1, e2, e3, sep=""), because it creates a new (temporary) result vector each time the operator is applied.

Value

Returns a character vector.

Examples

c('abc', '123', '\u0105\u0104') %stri+% letters[1:6]
'ID_' %stri+% 1:5

[Package stringi version 1.0-1 Index]