stri_escape_unicode {stringi}	R Documentation

Escape Unicode Code Points

Description

Escapes all Unicode (not ASCII-printable) code points.

Usage

stri_escape_unicode(str)

Arguments

str
character vector

Details

For non-printable and certain special (well-known, see also R man page Quotes) ASCII characters the following (also recognized in R) convention is used. We get \a, \b, \t, \n, \v, \f, \r, \", \', \\ or either \uXXXX (4 hex digits) or \UXXXXXXXX (8 hex digits) otherwise.

As usual, any input string is converted to Unicode before executing the escape process.

Value

Returns a character vector.

Author(s)

Marek Gagolewski and other contributors

See Also

The official online manual of stringi at https://stringi.gagolewski.com/

Gagolewski M., stringi: Fast and portable character string processing in R, Journal of Statistical Software 103(2), 2022, 1-59, doi:10.18637/jss.v103.i02

Other escape: stri_unescape_unicode()

Examples

stri_escape_unicode('a\u0105!')

[Package stringi version 1.7.12 Index]