stri_info {stringi}	R Documentation

Query Default Settings for stringi

Description

Gives the current default settings used by the ICU library.

Usage

stri_info(short = FALSE)

Arguments

short
logical; whether or not the results should be given in a concise form; defaults to TRUE

Value

If short is TRUE, then a single string providing information on the default character encoding, locale, and Unicode as well as ICU version is returned.

Otherwise, a list with the following components is returned:

Unicode.version – version of Unicode supported by the ICU library;

ICU.version – ICU library version used;

Locale – contains information on default locale, as returned by stri_locale_info;

Charset.internal – fixed at c('UTF-8', 'UTF-16');

Charset.native – information on the default encoding, as returned by stri_enc_info;

ICU.system – logical; TRUE indicates that the system ICU libs are used, otherwise ICU was built together with stringi;

ICU.UTF8 – logical; TRUE if the internal U_CHARSET_IS_UTF8 flag is defined and set.

Author(s)

Marek Gagolewski and other contributors

See Also

The official online manual of stringi at https://stringi.gagolewski.com/

Gagolewski M., stringi: Fast and portable character string processing in R, Journal of Statistical Software 103(2), 2022, 1-59, doi:10.18637/jss.v103.i02

[Package stringi version 1.7.12 Index]