stri_install_check {stringi}	R Documentation

Installation-Related Utilities [DEPRECATED]

Description

These functions are responsible for checking and guaranteeing that the ICU data library (icudt) is available and that stringi is ready to use.

These functions are deprecated and will no longer be available in future stringi releases.

Usage

stri_install_check(silent = FALSE)

stri_install_icudt(check = TRUE, outpath = NULL, inpath = NULL)

Arguments

silent
suppress diagnostic messages

check
enable stri_install_check() tests

outpath
path to install icudt to. If NULL, then file.path(path.package("stringi"), "libs") will be used.

inpath
path to search icudt archive in. If NULL, then only stringi mirror servers will be used. Mainly of interest to system admins and software developers.

Details

ICU makes use of a wide variety of data tables to provide many of its services. Examples include converter mapping tables, collation rules, transliteration rules, break iterator rules and dictionaries, and other locale data.

Without the ICU data library (icudt) many stringi features will not be available. icudt size is approx. 10-30 MB.

stri_install_check() tests whether some ICU services are available. If they are not, it is most likely due to unavailable ICU data library.

stri_install_icudt() downloads and installs the ICU data library specific for your platform (little/big-endian). The downloaded file will be decompressed into the directory where the package has been installed, see find.package, so make sure you have sufficient write permissions.

Value

These functions return a logical value, invisibly. TRUE denotes that the requested operation has been completed successfully.

References

ICU Data – ICU User Guide, http://userguide.icu-project.org/icudata

Examples

stri_install_check()

[Package stringi version 1.0-1 Index]