stri_locale_list {stringi}	R Documentation

List Available Locales

Description

Creates a character vector with all available locale identifies.

Usage

stri_locale_list()

Details

Note that some of the services may be unavailable in some locales. Querying for locale-specific services is always performed during the resource request.

See stringi-locale for more information.

Value

Returns a character vector with locale identifiers that are known to ICU.

Author(s)

Marek Gagolewski and other contributors

See Also

The official online manual of stringi at https://stringi.gagolewski.com/

Gagolewski M., stringi: Fast and portable character string processing in R, Journal of Statistical Software 103(2), 2022, 1-59, doi:10.18637/jss.v103.i02

Other locale_management: about_locale, stri_locale_info(), stri_locale_set()

Examples

stri_locale_list()

[Package stringi version 1.7.12 Index]