stri_read_raw {stringi}	R Documentation

Read Text File as Raw

Description

Reads a text file as-is, with no conversion or text line splitting.

Usage

stri_read_raw(con, fname = con)

Arguments

con
name of the output file or a connection object (opened in the binary mode)

fname
deprecated alias of con

Details

Once a text file is read into memory, encoding detection (see stri_enc_detect), conversion (see stri_encode), and/or splitting of text into lines (see stri_split_lines1) can be performed.

Value

Returns a vector of type raw.

Author(s)

Marek Gagolewski and other contributors

See Also

The official online manual of stringi at https://stringi.gagolewski.com/

Gagolewski M., stringi: Fast and portable character string processing in R, Journal of Statistical Software 103(2), 2022, 1-59, doi:10.18637/jss.v103.i02

Other files: stri_read_lines(), stri_write_lines()

[Package stringi version 1.7.12 Index]