stri_replace_rstr {stringi}	R Documentation

Convert gsub-Style Replacement Strings

Description

Converts a gsub-style replacement strings to those which can be used in stri_replace. In particular, $ becomes \$ and \1 becomes $1.

Usage

stri_replace_rstr(x)

Arguments

x
character vector

Value

Returns a character vector.

Author(s)

Marek Gagolewski and other contributors

See Also

The official online manual of stringi at https://stringi.gagolewski.com/

Gagolewski M., stringi: Fast and portable character string processing in R, Journal of Statistical Software 103(2), 2022, 1-59, doi: 10.18637/jss.v103.i02

Other search_replace: about_search, stri_replace_all(), stri_trim_both()

[Package stringi version 1.7.8 Index]