stri_enc_tonative {stringi}	R Documentation

Convert Strings To Native Encoding

Description

Converts character strings with declared encodings to Native encoding.

Usage

stri_enc_tonative(str)

Arguments

str
a character vector to be converted

Details

This function just calls stri_encode(str, NULL, NULL). Current native encoding can be read with stri_enc_get. Character strings declared to be in bytes encoding will fail here.

Note that if working in a UTF-8 environment, resulting strings will be marked with UTF-8 and not native, see stri_enc_mark.

Value

Returns a character vector.

See Also

Other encoding_conversion: stri_conv, stri_encode; stri_enc_fromutf32; stri_enc_toascii; stri_enc_toutf32; stri_enc_toutf8; stringi-encoding

[Package stringi version 1.0-1 Index]