stri_locale_info {stringi}	R Documentation

Query Given Locale

Description

Provides some basic information on a given locale identifier.

Usage

stri_locale_info(locale = NULL)

Arguments

locale
NULL or '' for default locale, or a single string with locale identifier.

Details

With this function you may obtain some basic information on any provided locale identifier, even if it is unsupported by ICU or if you pass a malformed locale identifier (the one that is not, e.g., of the form Language_Country). See stringi-locale for discussion.

This function does not do anything really complicated. In many cases it is similar to a call to as.list(stri_split_fixed(locale, '_', 3L)[[1]]), with locale case mapped. It may be used, however, to get insight on how ICU understands a given locale identifier.

Value

Returns a list with the following named character strings: Language, Country, Variant, and Name, being their underscore separated combination.

Author(s)

Marek Gagolewski and other contributors

See Also

The official online manual of stringi at https://stringi.gagolewski.com/

Other locale_management: about_locale, stri_locale_list(), stri_locale_set()

Examples

stri_locale_info('pl_PL')
stri_locale_info('Pl_pL') # the same result

[Package stringi version 1.7.6 Index]