stri_locale_list {stringi}	R Documentation

List Available Locales

Description

Creates a character vector with all available locale identifies.

Usage

stri_locale_list()

Details

Note that some of the services may be unavailable in some locales. Querying for locale-specific services is always performed during the resource request.

See stringi-locale for more information.

Value

Returns a character vector with locale identifiers that are known to ICU.

Author(s)

Marek Gagolewski and other contributors

See Also

The official online manual of stringi at https://stringi.gagolewski.com/

Other locale_management: about_locale, stri_locale_info(), stri_locale_set()

Examples

stri_locale_list()

[Package stringi version 1.7.6 Index]