stri_locale_set {stringi}	R Documentation

Set or Get Default Locale in stringi

Description

stri_locale_set changes the default locale for all the functions in the stringi package, i.e., establishes the meaning of the “NULL locale” argument of locale-sensitive functions. stri_locale_get gives the current default locale.

Usage

stri_locale_set(locale)

stri_locale_get()

Arguments

locale
single string of the form Language, Language_Country, or Language_Country_Variant, e.g., 'en_US', see stri_locale_list.

Details

See stringi-locale for more information on the effect of changing the default locale.

stri_locale_get is the same as stri_locale_info(NULL)$Name.

Value

stri_locale_set returns a string with previously used locale, invisibly.

stri_locale_get returns a string of the form Language, Language_Country, or Language_Country_Variant, e.g., 'en_US'.

Author(s)

Marek Gagolewski and other contributors

See Also

The official online manual of stringi at https://stringi.gagolewski.com/

Other locale_management: about_locale, stri_locale_info(), stri_locale_list()

Examples

## Not run: 
oldloc <- stri_locale_set('pt_BR')
# ... some locale-dependent operations
# ... note that you may always modify a locale per-call
# ... changing the default locale is convenient if you perform
# ... many operations
stri_locale_set(oldloc) # restore the previous default locale

## End(Not run)

[Package stringi version 1.7.6 Index]